requirements.txt
numpy
cython
pandas
pyyaml
# only for testing:
ddt
fitparse
